smoll steps

+++ Tabula/Infrastructure/DbModels/ItemDbModel.cs
@@ -4,7 +4,7 @@ public class ItemDbModel
 {
     public Guid Id { get; set; }
     public required string ProductName { get; set; }
-    public int Quantity { get; set; }
+    public ushort Quantity { get; set; }
     public bool Bought { get; set; }
     public Guid ShoppingListId { get; set; }
 }
